fix: rename held receive-policy transfer
