+++ tasks/tempo-v1/dataset.toml
@@ -22,8 +22,8 @@ name = "tempo-v1/faucet-funded-transfer"
 digest = "sha256:22cf474b1ead6304a02da0bf315797bc4f7383f6743fc4568192d0a73ca65f0b"
 
 [[tasks]]
-name = "tempo-v1/receive-policy-bounced-transfer"
-digest = "sha256:d70999941518d8dcbf02e8851e4dca8c9a985d5a6f74f888b57886d58546293f"
+name = "tempo-v1/receive-policy-held-transfer"
+digest = "sha256:7f03d2873c5c209477fac74b38505f39a22a2bcd61a273f9a8f549da29829616"
 
 [[tasks]]
 name = "tempo-v1/set-fee-token"
