+++ tasks/tempo-v1/receive-policy-held-transfer/environment/Dockerfile
